[fix]エラーは出なくなったが何も表示されない

+++ src/lib/api/auth.ts
@@ -15,11 +15,12 @@ export const signIn = (params: SignInParams) => {
 
 // サインアウト（ログアウト）
 export const signOut = () => {
+  if (!Cookies.get("_access_token") || !Cookies.get("_client") || !Cookies.get("_uid")) return
   return client.delete("auth/sign_out", {
     headers: {
-      "access-token": Cookies.get("_access_token"),
-      "client": Cookies.get("_client"),
-      "uid": Cookies.get("_uid")
+      "access-token": Cookies.get("_access_token")!,
+      "client": Cookies.get("_client")!,
+      "uid": Cookies.get("_uid")!
     }
   })
 }
@@ -29,9 +30,9 @@ export const getCurrentUser = () => {
   if (!Cookies.get("_access_token") || !Cookies.get("_client") || !Cookies.get("_uid")) return
   return client.get("/auth/sessions", {
     headers: {
-      "access-token": Cookies.get("_access_token"),
-      "client": Cookies.get("_client"),
-      "uid": Cookies.get("_uid")
+      "access-token": Cookies.get("_access_token")!,
+      "client": Cookies.get("_client")!,
+      "uid": Cookies.get("_uid")!
     }
   })
 }
